Time: 1 ms (93.22%), Space: 75.6 MB (24.21%) - LeetHub

+++ 1248-count-number-of-nice-subarrays/1248-count-number-of-nice-subarrays.cpp
@@ -6,9 +6,9 @@ public:
             if(nums[i]%2!=0) idx.push_back(i);
         }
         int cnt = 0;
-        for(int i = 0; i + k <= (int)idx.size();i++){
+        for(int i = 0; i + k <= idx.size();i++){
         int left = (i-1 < 0) ? (idx[i] + 1) : (idx[i] - idx[i-1]);
-        int right = (i+k < (int)idx.size()) ? (idx[i+k] - idx[i+k-1]):((int)nums.size() - idx[i+k-1]);
+        int right = (i+k < idx.size()) ? (idx[i+k] - idx[i+k-1]):(nums.size() - idx[i+k-1]);
         cnt += left * right;
         }
         return cnt;
